# server/src/docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: DefiScan APIs
  description: DefiScan APIs
  license:
    name: Github
    url: 'https://github.com/defichain/defiscan'
...
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/txs:
    get:
      tags:
        - Transactions
      summary: Get list blocks. Params limit & page & ((address & tx_account) | block | type)
      parameters:
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
        - name: address
          in: query
          required: false
          description: list tx of address, group with tx_account
          type: string
        - name: tx_account
          in: query
          required: false
          description: in|out, default = all, group with address
          type: string
        - name: block
          in: query
          required: false
          description: list tx of a block
          type: number
        - name: type
          in: query
          required: false
          description: signTxs|otherTxs|pending|all
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/txs/{hash}:
    get:
      tags:
        - Transactions
      summary: Get transaction detail
      parameters:
        - name: hash
          in: path
          required: true
          description: transaction hash
          type: string
      responses:
        200:
          description: OK
        404:
          description: Not Found
        500:
          description: Server Internal Error
...